adding boost reward tag
