+++ app/leaderboard/page.tsx
@@ -317,6 +317,7 @@ export default function Page() {
               reward={featuredQuest?.rewards_title}
               desc={featuredQuest?.desc}
               expiry={featuredQuest?.expiry_timestamp}
+              questId={featuredQuest?.id}
             />
           </div>
           <div className={styles.leaderboard_layout}>
