+++ components/quests/quest.tsx
@@ -1,10 +1,4 @@
-import React, {
-  FunctionComponent,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import React, { FunctionComponent, useMemo } from "react";
 import { useContext } from "react";
 import { QuestsContext } from "@context/QuestsProvider";
 import CheckIcon from "@components/UI/iconsComponents/icons/checkIcon";
@@ -12,10 +6,7 @@ import UnavailableIcon from "@components/UI/iconsComponents/icons/unavailableIco
 import styles from "@styles/quests.module.css";
 import { CDNImg } from "@components/cdn/image";
 import QuestCard from "./questCard";
-import { getBoosts } from "@services/apiService";
-import TokenSymbol from "@components/quest-boost/TokenSymbol";
-import { TOKEN_DECIMAL_MAP } from "@utils/constants";
-import { getTokenName } from "@utils/tokenService";
+import BoostReward from "./boostReward";
 
 type QuestProps = {
   onClick: () => void;
@@ -36,42 +27,11 @@ const Quest: FunctionComponent<QuestProps> = ({
   id,
   expired,
 }) => {
-  const { completedQuestIds, boostedQuests } = useContext(QuestsContext);
+  const { completedQuestIds } = useContext(QuestsContext);
   const isCompleted = useMemo(
     () => completedQuestIds.includes(id),
     [id, completedQuestIds]
   );
-  const [boost, setBoost] = useState<Boost>();
-  const [isQuestBoosted, setIsQuestBoosted] = useState<boolean>(false);
-
-  const checkIfBoostedQuest = useCallback(async () => {
-    if (!boostedQuests) return;
-    if (boostedQuests.length > 0 && boostedQuests.includes(id))
-      setIsQuestBoosted(true);
-  }, []);
-
-  const fetchBoosts = useCallback(async (id: string) => {
-    try {
-      const response = await getBoosts();
-      if (!response) return;
-      const boost = response.find((b: Boost) =>
-        b.quests.includes(parseInt(id))
-      );
-      if (!boost) return;
-      setBoost(boost);
-    } catch (err) {
-      console.log("Error while fetching boost by id", err);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!isQuestBoosted) return;
-    fetchBoosts(id.toString());
-  }, [isQuestBoosted, id]);
-
-  useEffect(() => {
-    checkIfBoostedQuest();
-  }, []);
 
   return (
     <QuestCard
@@ -106,15 +66,7 @@ const Quest: FunctionComponent<QuestProps> = ({
             </>
           )}
         </div>
-        {boost && isQuestBoosted ? (
-          <div
-            className={styles.issuer}
-            style={{ gap: 0, padding: "8px 16px" }}
-          >
-            <TokenSymbol tokenAddress={boost?.token} />
-            <p className="text-white ml-2">{boost?.amount}</p>
-          </div>
-        ) : null}
+        <BoostReward questId={id} />
       </div>
     </QuestCard>
   );
